// src/app/models/car-information.model.ts
export class FullCarInformation {
  id: string;
  total_score: string;
  car_image: string;
  name: string;
  price: number;
  type: string;
  car_type: string;
  imagepath: string;
  carrying_people: string;
  engine: string;
  torque_nm: string;
  acceleration_0_100: string;
  consumption_l_100km: string;
  verdict: string;
  engine_type: string;
  power_kw: string;
}